Another choice tabs width
There is a way to separate the favicon from the title (inside the tab) ? I ignore.
It's might be better to purpose a second choice. Because you know the people are able to identify (associate, etc) a website just with a logo (favicon, picture, etc).
Another reason it will help because the space screen is very limited (that's true for a laptop, etc). Sometimes the people when to open a lot of website at once (let's say twenty, etc)