Make hovering dark tabs less distracting by making the tab separator fixed
When using the alternative appearance for background tabs, the animation
of the "after" separator on "hover" can be a little distracting, maybe
because the animation was designed for the normal appearance.

Set the margins of the separator to a fixed value, this way the
separator does not animate at all on "hover" and looks more consistent.